feat(cli): restructure CLI into new hierarchy
