+++ tests/integration.rs
@@ -51,6 +51,7 @@ pub fn keepldr_exec<'a>(bin: impl Into<PathBuf>, input: impl Into<Option<&'a [u8
     let bin: PathBuf = bin.into();
     let mut child = Command::new(KEEP_BIN)
         .current_dir(CRATE)
+        .arg("unstable")
         .arg("exec")
         .arg(&bin)
         .stdin(Stdio::piped())
